Separated search function out into separate component to make it easier to test and easier to move when switching to mobile view
