+++ stock-tracker/src/assets/styles/index.css
@@ -299,7 +299,6 @@ h1 {
 }
 
 .news {
-  /* margin-bottom: 12px; */
   flex: 1 1 auto;
 }
