+++ stock-tracker/src/components/Header.js
@@ -1,84 +1,16 @@
-import React, { useState, useCallback, useEffect, useRef } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { addSymbolAction, addSearchInputAction } from "../store/actions";
+import React from "react";
+import { useSelector } from "react-redux";
 import logo from "../assets/logo.png";
 import { Icon } from "antd";
 import "../assets/styles/Header.css";
-import sun from "../assets/sun.png"
-import moon from "../assets/moon.png"
+import sun from "../assets/sun.png";
+import moon from "../assets/moon.png";
+import Search from "./Search";
 const moment = require("moment");
 
 const Header = () => {
-  const [symbol, setSymbol] = useState("");
-  const [open, setOpen] = useState(false);
-  const dropdownRef = useRef(null);
-  const searchRef = useRef(null);
-  const dispatch = useDispatch();
-  const addSymbol = useCallback(symbol => dispatch(addSymbolAction(symbol)), [
-    dispatch
-  ]);
-  const addSearchInput = useCallback(
-    searchInput => dispatch(addSearchInputAction(searchInput)),
-    [dispatch]
-  );
   const response = useSelector(state => state.response);
   const overview = useSelector(state => state.companyOverview);
-  const suggestions = useSelector(state => state.suggestions);
-  const handleSubmit = e => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      if (symbol.trim() === "") return;
-      addSymbol(symbol);
-      setOpen(false);
-      setSymbol("");
-    }
-  };
-
-  const onChange = e => {
-    setSymbol(e.target.value);
-    addSearchInput(e.target.value);
-  };
-
-  const onClick = e => {
-    addSymbol(e.target.id);
-    setOpen(false);
-    setSymbol("");
-  };
-
-  const handleBlur = () => {
-    requestAnimationFrame(() => {
-      if (
-        !dropdownRef.current.contains(document.activeElement) &&
-        !searchRef.current.contains(document.activeElement)
-      ) {
-        setOpen(false);
-      } else {
-        searchRef.current.focus();
-      }
-    });
-  };
-
-  useEffect(() => {
-    setOpen(suggestions !== 0);
-  }, [suggestions]);
-
-  const suggestionItems =
-    suggestions.length > 0
-      ? suggestions.map(data => {
-          return (
-            <li
-              className="search_display__suggestion_list__item"
-              onClick={onClick}
-              id={data.symbol}
-              key={data.symbol}
-            >
-              <p className="search_display__suggestion_list__item__symbol">{`(${data.symbol}) `}</p>
-              {` ${data.name}`}
-            </li>
-          );
-        })
-      : "";
-
   const changeNo =
     response.change === 0
       ? "0"
@@ -94,12 +26,11 @@ const Header = () => {
     ? `Real time price as of ${response.latestTime} ${formatedTime}`
     : "";
 
-    const statusIcon = response
+  const statusIcon = response
     ? marketStatus === "Market Open"
       ? sun
       : moon
-    : ""
-
+    : "";
   return (
     <div className="header">
       <img className="header__logo" alt="logo" src={logo} />
@@ -109,29 +40,7 @@ const Header = () => {
         <button className="header__btns__btn">WATCHLIST</button>
       </div>
       <div className="search_display">
-        <h5>
-          <Icon className="search_display__icon" type="search" />
-        </h5>
-        <input
-          className="search_display__search_bar"
-          placeholder={
-            response ? `${response.companyName}(${response.symbol})` : ""
-          }
-          type="text"
-          value={symbol}
-          onChange={onChange}
-          onKeyPress={handleSubmit}
-          onBlur={handleBlur}
-          ref={searchRef}
-        />
-        <ul
-          ref={dropdownRef}
-          tabIndex="-1"
-          className="search_display__suggestion_list"
-          style={{ display: open ? "block" : "none" }}
-        >
-          {suggestionItems}
-        </ul>
+        <Search />
       </div>
       <div className="price_display">
         <p className="price_display__small_icon">{response ? "$" : ""}</p>
@@ -181,7 +90,7 @@ const Header = () => {
       </div>
       <div className="market_status_display">
         <p className="market_status_display__real_time">{realTimeDisplay}</p>
-        <img className="market_status__icon" src={statusIcon}/>
+        <img className="market_status__icon" src={statusIcon} />
         <p className="market_status_display__real_time__status">
           {marketStatus}
         </p>
